git
Enum Commands

pub(crate) enum Commands {
    Init {
        initial_branch: Option<String>,
        common: CommonArgs,
    },
    Add {
        files: Vec<String>,
        common: CommonArgs,
    },
    Rm {
        recursive: bool,
        cached: bool,
        files: Vec<String>,
        common: CommonArgs,
    },
    Commit {
        message: String,
        common: CommonArgs,
    },
    Branch {
        name: Option<Vec<String>>,
        delete: bool,
        common: CommonArgs,
    },
    Checkout {
        target: String,
        force: bool,
        branch: bool,
        common: CommonArgs,
    },
    Merge {
        branch: String,
        force: bool,
        common: CommonArgs,
    },
    Status {
        common: CommonArgs,
    },
    Log {
        common: CommonArgs,
    },
}

Variants

Init

Initialize a new Git repository

Fields

initial_branch: Option<String>

Use the specified name for the initial branch in the newly created repository.

common: CommonArgs

Add

Add file(s) to the index (staging area)

Fields

files: Vec<String>

List of files to add

common: CommonArgs

Rm

Remove file(s) from the index

Fields

recursive: bool

Allow recursive removal when a leading directory name is given.

cached: bool

Only remove cached files (in staging area).

files: Vec<String>

List of files to remove

common: CommonArgs

Commit

Commit staged changes

Fields

message: String

Commit message

common: CommonArgs

Branch

Create, list, or delete branches

Fields

name: Option<Vec<String>>

Branch name (if omitted, list branches)

delete: bool

Delete the specified branch

common: CommonArgs

Checkout

Switch to another branch

Fields

target: String

Target branch name

force: bool

Force checkout even there are uncommited changes.

branch: bool

Create a new branch and checkout

common: CommonArgs

Merge

Merge the specified branch into the current one

Fields

branch: String

Branch to merge from

force: bool

Force merge even there are uncommited changes.

common: CommonArgs

Status

Displays paths that have differences between the index file and the current HEAD commit

Fields

common: CommonArgs

Log

Displays the log of commits

Fields

common: CommonArgs